Fixed a bug which caused the program to crash when modifying detector foil layers.

+++ dialogs/simulation/layer_properties.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 """
 Created on 28.2.2018
-Updated on 20.8.2018
+Updated on 21.8.2018
 
 Potku is a graphical user interface for analyzation and
 visualization of measurement data collected from a ToF-ERD
@@ -324,14 +324,15 @@ class LayerPropertiesDialog(QtWidgets.QDialog):
                     if elem_sim.controls:
                         elem_sim.controls.reset_controls()
 
-                for energy_spectra in self.tab.energy_spectrum_widgets:
-                    self.tab.del_widget(energy_spectra)
-                    save_file_path = os.path.join(
-                        self.tab.simulation.directory,
-                        energy_spectra.save_file)
-                    if os.path.exists(save_file_path):
-                        os.remove(save_file_path)
-                self.tab.energy_spectrum_widgets = []
+                if self.tab:
+                    for energy_spectra in self.tab.energy_spectrum_widgets:
+                        self.tab.del_widget(energy_spectra)
+                        save_file_path = os.path.join(
+                            self.tab.simulation.directory,
+                            energy_spectra.save_file)
+                        if os.path.exists(save_file_path):
+                            os.remove(save_file_path)
+                    self.tab.energy_spectrum_widgets = []
 
                 for elem_sim in simulations_run:
                     for recoil in elem_sim.recoil_elements:
@@ -377,14 +378,15 @@ class LayerPropertiesDialog(QtWidgets.QDialog):
                     if elem_sim.controls:
                         elem_sim.controls.reset_controls()
 
-                for energy_spectra in self.tab.energy_spectrum_widgets:
-                    self.tab.del_widget(energy_spectra)
-                    save_file_path = os.path.join(
-                        self.tab.simulation.directory,
-                        energy_spectra.save_file)
-                    if os.path.exists(save_file_path):
-                        os.remove(save_file_path)
-                self.tab.energy_spectrum_widgets = []
+                if self.tab:
+                    for energy_spectra in self.tab.energy_spectrum_widgets:
+                        self.tab.del_widget(energy_spectra)
+                        save_file_path = os.path.join(
+                            self.tab.simulation.directory,
+                            energy_spectra.save_file)
+                        if os.path.exists(save_file_path):
+                            os.remove(save_file_path)
+                    self.tab.energy_spectrum_widgets = []
 
         elif simulations_run:
             reply = QtWidgets.QMessageBox.question(
@@ -412,14 +414,15 @@ class LayerPropertiesDialog(QtWidgets.QDialog):
                     if elem_sim.controls:
                         elem_sim.controls.reset_controls()
 
-                for energy_spectra in self.tab.energy_spectrum_widgets:
-                    self.tab.del_widget(energy_spectra)
-                    save_file_path = os.path.join(
-                        self.tab.simulation.directory,
-                        energy_spectra.save_file)
-                    if os.path.exists(save_file_path):
-                        os.remove(save_file_path)
-                self.tab.energy_spectrum_widgets = []
+                if self.tab:
+                    for energy_spectra in self.tab.energy_spectrum_widgets:
+                        self.tab.del_widget(energy_spectra)
+                        save_file_path = os.path.join(
+                            self.tab.simulation.directory,
+                            energy_spectra.save_file)
+                        if os.path.exists(save_file_path):
+                            os.remove(save_file_path)
+                    self.tab.energy_spectrum_widgets = []
 
         name = self.__ui.nameEdit.text()
         thickness = self.__ui.thicknessEdit.value()
